Send back NFT to original owner
When the "signs and executes a tx block"
test was done, the recipient held the
NFT, making the tests non reproducible
back-to-back. We would have to manually
send back every time the NFT back to
its' original owner (the admin).

With this commit, the NFT is sent back
to the original owner after the test
completes and everything runs smoothly.

IMPORTANT NOTE-
The NFT won't be sent back to the og owner
if the test fails in prior steps!