Test multiple transactions (SYNC)
So far the code was tested only
for a single txb to be handled by
the executorServiceHandler.

Now, we introduced 3 txbs to be
executed by it. BUT the code
gets executed synchronously!
To take advantage of our pool architecture,
we need to use multiple workers asynchronously
and handle the requests (txbs).