Lazy load objects instead of storing them in main pool
This is a huge performance improvement.

- Old version:
So far we have been fetching with multiple requests
all the objects of the admin account and storing them
to the mainPool.

- New version:
This time we lazily load the objects only when needed.

The main pool should only keep a batch of objects.
During split, if the split strategy is successful (i.e.
all the objects that it needs are included in the new pool)
the split function will short circuit (stop) and return
the new pool.

If the current batch that the main pool contains does
not fulfill the split strategy, then it moves on by
loading the next batch of objects.

We implement the above using the generator pattern
since it is easier to think about.

--- TESTS ---
Tests had to be updated. Mostly tests that used the
mainPool to execute transactions. This had to change,
since as we mentioned above mainPool only contains a
specific subset of objects (which might not be enough
to execute the transaction).
We always have to split the main pool and then execute
the transaction using the new pool.
Remember that the splitStrategy is the one that defines
what objects and coins are needed to be included in the
new pool in order for the transaction to succeed.