Fix - save split predicate rejected objects to mainPool
So far rejected objects from the predicate were ignored.
We were iterating the generator from the beginning
each time we did a new pool split.

Now we keep a state (using the cursor) of the queried objects
and a memory (this._objects) of the objects we used.
Whenever we split, the predicate iterates the this._objects
(i.e. the mainpool objects -memory-) and then if it is unsuccessful,
more objects are queried and added to the mainPool with fetchObjects.
The predicate re-iterates the data.