Simplify isCoin function
So far isCoin would use a regex
to determine if a coin is of a
specific type.

That would not be robust enough
because there would be many cases
breaking it such as:

1. When checking the marker type,
you again need to check the fully
resolved type, not just the type name.

2. When dealing with addresses, need to
be careful about canonicalisation (0x2
and 0x02 should be considered the same address).

On the other hand, passing the objectType
(i.e. '0x2::coin::Coin<0x2::sui::SUI>' instead of 'SUI')
does not leave any space for errors.
Other object types should be passed in the same way.