Include balance in PoolObject(s) and update SplitStrategy
The DefaultSplitStrategy has now changed so that
each new pool that is being created will contain
a group of coins where the sum of their balances
satisfies a minimum threshold.

To do this we each PoolObject needs to contain
an optional balance field. The balance field
will be not-null only for gas coin objects.

Notes:
- Updated unit tests to assert balance changes
- Fixed new gasCoins calculation: gasCoins property
was not updating after a pool was split. Meaning that
even though some gas coins were moved to the new pool,
they would still show in the initial pool's gasCoins.