Change assertions of test "uses only the pool s coins for gas"
So far we have been comparing the exact gasBill
with the exact balance change of the pool that
executed the transaction. This is not necessary and
imprecise, resulting to tests failing sometimes (could
be also be due wrong formula of gas bill calculation).

We don't need to check the exact amount paid
by the new pool. We only need to check if
the mainPool was not used to pay for gas
and only the new pool did.