Add support for sponsored transactions
A new lambda parameter is added to the
ExecutorServiceHandler.execute() that
is called before the transaction block
execution. This lambda generates a pair of
signed transaction blocks (signed from the
sender and the sponsor).

Its purpose it to be used as a way of sponsoring
a TransactionBlockWithLambda - i.e., a transaction
block that uses placeholders for unknown object ids
such as AdminCaps.

Additional changes:
- A new test that executes a sponsored transaction.
- Code refactoring.
- New split strategy that only fetched AdminCap objects
(without selecting coins, as they are not needed for
sponsored transactions).